fixed bug maybe

+++ index.js
@@ -150,7 +150,7 @@ app.post('/login', async (req, res) => {
 });
 
 function createAuthToken(username, res) {
-    const expiration = (process.env.JWT_EXPIRATION || '15') + 'min';
+    const expiration = (process.env.JWT_EXPIRATION || '30') + 'min';
     const accessToken = jwt.sign({ username }, process.env.ACCESS_TOKEN_SECRET, { expiresIn: expiration });
 
     console.log('Creating auth token for', username);
@@ -414,7 +414,7 @@ app.post('/manager/proposed-problems/:id', authenticateToken, isManager, async (
         if (difficulty <= 0) return res.status(400).json({ message: 'difficulty must be positive' });
 
         let result = await query(`INSERT INTO problems (name, description, content, time_created, testable, answer, creator_id, difficulty) VALUES ($1, $2, $3, $4, $5, $6, $7, $8)`, [name, description, content, proposedProblem.timeCreated, testable, proposedProblem.answer, proposedProblem.creatorId, difficulty]);
-        if (result instanceof Error) return res.status(500).json({ errorCode: result.code, message: 'Failed to add to problems table' });
+        if (result instanceof Error || !result) return res.status(500).json({ errorCode: result.code, message: 'Failed to add to problems table' });
         problems = await update('problems');
 
         result = await query(`DELETE FROM proposed_problems WHERE id = $1`, [req.params.id]);
